% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{convert_type}
\alias{convert_type}
\title{Convert R type specification to JSON Schema type}
\usage{
convert_type(type_spec)
}
\arguments{
\item{type_spec}{The type specification from the schema}
}
\value{
A list representing the JSON Schema type
}
\description{
Convert R type specification to JSON Schema type
}
\keyword{internal}
