% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{create_JsonSchema}
\alias{create_JsonSchema}
\title{Create a new JsonSchema object}
\usage{
create_JsonSchema(schema, json_schema_str)
}
\arguments{
\item{schema}{An R list representing the desired output structure.}

\item{json_schema_str}{The JSON schema as a string.}
}
\value{
A JsonSchema S7 object.
}
\description{
Helper function to create a new JsonSchema object.
}
