% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_multi_gap}
\alias{extract_multi_gap}
\title{Extract multiple regions of an alignment concatenated}
\usage{
extract_multi_gap(query, start = 1, end = 1)
}
\arguments{
\item{query}{A DNAStringSet Object. It must be set.}

\item{start}{A vector indicating each start position for each region. Defaults to 1.}

\item{end}{A vector indicating each end position for each region. Defaults to 1.}
}
\value{
The function returns a DNAStringSet of the selected regions of the alignment according to the
Start and End parameters.
}
\description{
Use this function to extract specific regions of interest from
an alignment concatenated sequentially.
}
\examples{
extract_multi_gap(c("acccataaaa","agtaatagat"),
                 start=2,
                 end=8)

extract_multi_gap(c("acccataaaa","agtaatagat"),
                 start=2,
                 end=8)
}
