% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_seq}
\alias{extract_seq}
\title{Extract a single region of the alignment}
\usage{
extract_seq(query, start, end)
}
\arguments{
\item{query}{A DNAStringSet Object. It must be set.}

\item{start}{A vector indicating each start position. Defaults to 1.}

\item{end}{A vector indicating each end position. Defaults to 1.}
}
\value{
The function returns a DNAStringSet of the selected region of the alignment according to the
Start and End parameters.
}
\description{
Extract a single region of the alignment
}
\examples{
extract_seq("acccataaaa",start=2,end=8)
extract_seq("acccataaaa",1,4)
}
