% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{JsonSchema}
\alias{JsonSchema}
\title{S7 Class for JSON Schema}
\usage{
JsonSchema(schema = list(), json_schema_str = character(0))
}
\arguments{
\item{schema}{An R list representing the desired output structure.}

\item{json_schema_str}{The JSON schema as a string.}
}
\description{
An S7 class to represent a JSON schema, including its R list definition
and its string representation.
}
\section{Slots}{

\describe{
\item{\code{schema}}{An R list representing the desired output structure.}

\item{\code{json_schema_str}}{The JSON schema as a string.}
}}

