\name{ParseDescriptionFiles}
\alias{ParseDescriptionFiles}
\encoding{UTF-8}
\title{ParseDescriptionFiles}
\description{
  Parses DESCRIPTION files from R packages.
}

\usage{
ParseDescriptionFiles(datadir)
}
\arguments{
  \item{datadir}{Directory where to store data}
}
\details{
  Parses DESCRIPTION file content, dependencies, authors and dates.
}

\author{
  Maëlick Claes <himself@maelick.net>
}

\keyword{description}
\keyword{dependencies}
\keyword{maintainers}
\keyword{cran}
\keyword{github}
\keyword{package}