\name{Extract}
\alias{Extract}
\encoding{UTF-8}
\title{Extract}
\description{
  Extract R data
}

\usage{
Extract(datadir)
}
\arguments{
  \item{datadir}{Directory where to store data}
}
\details{
  Extracts all informations from Github and CRAN packages.
}

\author{
  Maëlick Claes <himself@maelick.net>
}

\keyword{extract}
\keyword{cran}
\keyword{github}
\keyword{broken}
\keyword{package}