\name{ParseNamespaceFiles}
\alias{ParseNamespaceFiles}
\encoding{UTF-8}
\title{ParseNamespaceFiles}
\description{
  Parses NAMESPACE files from R packages.
}

\usage{
ParseNamespaceFiles(datadir)
}
\arguments{
  \item{datadir}{Directory where to store data}
}
\details{
  Parses NAMESPACE file content, dependencies, authors and dates.
}

\author{
  Maëlick Claes <himself@maelick.net>
}

\keyword{namespace}
\keyword{dependencies}
\keyword{import}
\keyword{export}
\keyword{cran}
\keyword{github}
\keyword{package}